chore(refactoring): made a new modal to delete the user
