--- components/directory/DeleteUserModal.js
@@ -1,128 +1,0 @@
-import { Dialog, Transition } from "@headlessui/react";
-import { Fragment, useRef } from "react";
-
-import { ExclamationTriangleIcon } from "@heroicons/react/24/outline";
-import PropTypes from "prop-types";
-
-export const DeleteUserModal = ({
-  selectedUser,
-  handleDeleteUser,
-  openModal,
-  handleCloseModal,
-  isSubmitting,
-}) => {
-  const cancelButtonRef = useRef(null);
-
-  return (
-    <Transition.Root show={openModal} as={Fragment}>
-      <Dialog
-        as="div"
-        className="relative z-10"
-        initialFocus={cancelButtonRef}
-        onClose={handleCloseModal}
-      >
-        <Transition.Child
-          as={Fragment}
-          enter="ease-out duration-300"
-          enterFrom="opacity-0"
-          enterTo="opacity-100"
-          leave="ease-in duration-200"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0"
-        >
-          <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
-        </Transition.Child>
-
-        <div className="fixed inset-0 z-10 overflow-y-auto">
-          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-            <Transition.Child
-              as={Fragment}
-              enter="ease-out duration-300"
-              enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-              enterTo="opacity-100 translate-y-0 sm:scale-100"
-              leave="ease-in duration-200"
-              leaveFrom="opacity-100 translate-y-0 sm:scale-100"
-              leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-            >
-              <Dialog.Panel className="relative transform overflow-hidden rounded-lg bg-white px-4 pt-5 pb-4 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6">
-                <div className="sm:flex sm:items-start">
-                  <div className="mx-auto flex h-12 w-12 flex-shrink-0 items-center justify-center rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
-                    <ExclamationTriangleIcon
-                      className="h-6 w-6 text-red-600"
-                      aria-hidden="true"
-                    />
-                  </div>
-                  <div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
-                    <Dialog.Title
-                      as="h3"
-                      className="text-base font-semibold leading-6 text-gray-900"
-                    >
-                      Eliminar Usuario
-                    </Dialog.Title>
-                    <div className="mt-2">
-                      <p className="text-sm text-gray-500">
-                        Usted esta a punto de eliminar al usuario{" "}
-                        <span className="text-red-500 font-semibold">
-                          {selectedUser?.fullname}
-                        </span>
-                        . Esta seguro que desea eliminarlo? Esta accion es
-                        irreversible.
-                      </p>
-                    </div>
-                  </div>
-                </div>
-                <div className="mt-5 sm:mt-4 sm:flex sm:flex-row-reverse">
-                  <button
-                    disabled={isSubmitting || false}
-                    type="button"
-                    className={
-                      isSubmitting
-                        ? "inline-flex w-full justify-center rounded-md bg-gray-300 px-3 py-2 text-sm font-semibold text-gray-400 sm:ml-3 sm:w-auto"
-                        : "inline-flex w-full justify-center rounded-md bg-red-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-red-500 sm:ml-3 sm:w-auto"
-                    }
-                    onClick={handleDeleteUser}
-                  >
-                    Si, eliminar
-                  </button>
-                  <button
-                    type="button"
-                    className="mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50 sm:mt-0 sm:w-auto"
-                    onClick={handleCloseModal}
-                    ref={cancelButtonRef}
-                  >
-                    Cancelar
-                  </button>
-                </div>
-              </Dialog.Panel>
-            </Transition.Child>
-          </div>
-        </div>
-      </Dialog>
-    </Transition.Root>
-  );
-};
-
-DeleteUserModal.propTypes = {
-  selectedUser: PropTypes.shape({
-    active: PropTypes.bool,
-    color: PropTypes.string,
-    company: PropTypes.string,
-    email: PropTypes.string,
-    firstname: PropTypes.string,
-    fullname: PropTypes.string.isRequired,
-    gender: PropTypes.string,
-    lastname: PropTypes.string,
-    notes: PropTypes.string,
-    phone: PropTypes.string.isRequired,
-    placeholder: PropTypes.string,
-    second_lastname: PropTypes.string,
-    second_name: PropTypes.string,
-    secondary_phone: PropTypes.string,
-    third_phone: PropTypes.string,
-    uniqueId: PropTypes.string,
-  }),
-  handleDeleteUser: PropTypes.func.isRequired,
-  openModal: PropTypes.bool.isRequired,
-  handleCloseModal: PropTypes.func.isRequired,
-  isSubmitting: PropTypes.bool.isRequired,
-};
